convert project tabs into separate pages
